+++ tests/functional/test_help.py
@@ -5,8 +5,7 @@ import pytest
 from pip._internal.cli.status_codes import ERROR, SUCCESS
 from pip._internal.commands import commands_dict, create_command
 from pip._internal.exceptions import CommandError
-from tests.conftest import InMemoryPip
-from tests.lib import PipTestEnvironment
+from tests.lib import InMemoryPip, PipTestEnvironment
 
 
 def test_run_method_should_return_success_when_finds_command_name() -> None:
