+++ tests/functional/test_inspect.py
@@ -2,8 +2,7 @@ import json
 
 import pytest
 
-from tests.conftest import ScriptFactory
-from tests.lib import PipTestEnvironment, TestData
+from tests.lib import PipTestEnvironment, ScriptFactory, TestData
 
 
 @pytest.fixture(scope="session")
